refactor(DevicesView): extract components
